ADD: support show play time in title
支持在标题栏中显示游戏时间和角色年龄

+++ gmud-simple/src/cn/fmsoft/lnx/gmud/simple/core/Video.java
@@ -267,10 +267,10 @@ class Video {
 		path.moveTo(x, y);
 		if ((type & 1) == 0) {
 			path.lineTo(x + w, y);
-			path.lineTo(x + w / 2, y + h);
+			path.lineTo(x + (float) w / 2, y + h);
 		} else {
 			path.lineTo(x, y + h);
-			path.lineTo(x + w, y + h / 2);
+			path.lineTo(x + w, y + (float) h / 2);
 		}
 		path.close();
